+++ src/router/index.tsx
@@ -1,8 +1,8 @@
-import { createBrowserRouter } from "react-router-dom";
+import { createHashRouter } from "react-router-dom";
 
 import routes from "./routes";
 
 //可传第二个参数，配置base路径，例如{ basename: "/app"}
-const router = createBrowserRouter(routes);
+const router = createHashRouter(routes);
 
 export default router;